requirements.txt
aiohttp>=3.8.0
openai>=1.0.0
pytesseract>=0.3.10
Pillow>=10.0.0
python-dotenv>=1.0.0
